refresh token, jwt token, migrations, protecting routes

+++ resources/item.py
@@ -6,17 +6,23 @@ from resources.schemas import ItemSchema, ItemUpdateSchema
 from models.item import ItemModel
 from db import db
 from sqlalchemy.exc import SQLAlchemyError
+from flask_jwt_extended import jwt_required, get_jwt
 
 blp = Blueprint("Items", __name__, description="Operations on items")
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            return {"message": "Admin privilege required."}, 401
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -39,13 +45,16 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True)) #can return multiple items
     def get(self):
         # return {"items": list(items.values())} # pre @blp.response
         # return items.values()
         return ItemModel.query.all()
     
-    @blp.arguments(ItemSchema)    
+    @jwt_required(fresh=True)
+    @blp.arguments(ItemSchema)
+    @blp.response(201, ItemSchema)    
     def post(self, item_data):
         item = ItemModel(**item_data)
